feat: generate strings
This commit brings in the ability to generate strings. It builds a
character set first, a set of acceptable values and just picks chars
from that set randomly to generate strings. This description is a bit
high-level and the actual implemenation utilizes the code point values
of the chars instead of the chars themselves.